requirements.txt
python-dotenv==1.0.1
telethon>=1.42.0,<2.0
kiteconnect==5.0.1
pandas==2.2.2
openpyxl==3.1.5
numpy==1.26.4
yfinance>=0.2.50
requests==2.32.3
